+++ R/ggmsa.R
@@ -2,7 +2,8 @@
 ##'
 ##'
 ##' @title ggmsa
-##' @param fasta Aligned FASTA format file for representing either nucleotide sequences or peptide sequences.
+##' @param msa Multiple aligned sequence file or object for
+##' representing either nucleotide sequences or peptide sequences.
 ##' @param start Start position to plot, If font=NULL, only the background frame is drawn, and no character.
 ##' @param end End position to plot,If font=NULL, only the background frame is drawn, and no character.
 ##' @param font Character font, Defaults is 'helvetica_regular'.
@@ -26,11 +27,10 @@
 ##' ggmsa(f, 164, 213, color="Chemistry_AA")
 ##' @export
 ##' @author guangchuang yu
-ggmsa <- function(fasta, start=NULL, end=NULL, font = "helvetica_regular", color = "Clustal") {
-    data <- msa_data(fasta, start = start, end = end,
-                     color = color, font = font)
+ggmsa <- function(msa, start=NULL, end=NULL, font = "helvetica_regular", color = "Clustal") {
+    data <- tidy_msa(msa, start = start, end = end)
 
-    ggplot() + geom_msa(data) + 
+    ggplot() + geom_msa(data, font = font, color = color) + 
         theme_minimal() + xlab(NULL) + ylab(NULL) +
         theme(legend.position='none') + coord_fixed()
 }